Refactor API types into centralized infrastructure types module
